+++ python/ray/rllib/__init__.py
@@ -9,7 +9,7 @@ from ray.tune.registry import register_trainable
 
 def _register_all():
     for key in ["PPO", "ES", "DQN", "APEX", "A3C", "BC", "PG", "DDPG",
-                "DDPG2", "APEX_DDPG2", "__fake", "__sigmoid_fake_data",
+                "DDPG2", "APEX_DDPG", "__fake", "__sigmoid_fake_data",
                 "__parameter_tuning"]:
         from ray.rllib.agent import get_agent_class
         register_trainable(key, get_agent_class(key))
