+++ python/ray/rllib/ddpg/README.md
@@ -1,0 +1,1 @@
+Implementation of deep deterministic policy gradients (https://arxiv.org/abs/1509.02971), including an Ape-X variant.
