+++ python/ray/rllib/tuned_examples/mountaincarcontinuous-apex-ddpg.yaml
@@ -1,12 +1,7 @@
 # This can be expected to reach 90 reward within ~1.5-2.5m timesteps / ~150-250 seconds on a K40 GPU
-mountaincarcontinuous-apex-ddpg-2:
+mountaincarcontinuous-apex-ddpg:
     env: MountainCarContinuous-v0
-    run: APEX_DDPG2
-    trial_resources:
-        cpu: 1
-        gpu: 1
-        extra_cpu:
-            eval: 4 + spec.config.num_workers
+    run: APEX_DDPG
     stop:
         episode_reward_mean: 90
     config:
